+++ controllers/home-routes.js
@@ -1,10 +1,6 @@
 const router = require('express').Router();
 const { User } = require('../models/');
 const withAuth = require('../utils/auth')
-const cors = require('cors');
-
-
-
 
 
 
@@ -14,19 +10,6 @@ router.get('/', (req, res) => {
   res.render('homepage', { loggedIn: req.session.loggedIn });
 });
 
-const corsOptions = {
-  origin: '*',
-  methods: ['GET','POST','OPTIONS'],
-  credentials: true,
-  allowedHeaders: ['Content-Type', 'authorization-bearer', 'Authorization'],
-  optionsSuccessfulStatue: 204,
-  optionsSuccessStatus: 200
-}
-
-router.options('*', cors(corsOptions));
-router.use(function(req,res, next){
-  res.header("Access-Control-Allow-Headers", "Origin, authorization-bearer, Authorization");
-});
 // Saved-destination route
 router.get('/dashboard', withAuth, async (req, res) => {
   try {
@@ -37,7 +20,7 @@ router.get('/dashboard', withAuth, async (req, res) => {
   }
 })
 
-
+// 
 router.post('/create', async (req, res) => {
   try {
     const dbUser = await User.create({
@@ -51,6 +34,7 @@ router.post('/create', async (req, res) => {
   }
 })
 
+// 
 router.get('/get', async (req, res) => {
   try {
     const dbUserData = await User.findAll({
@@ -74,6 +58,7 @@ router.get('/get', async (req, res) => {
   }
 })
 
+// 
 router.get('/get', async (req, res) => {
   try {
     const db = mysql.createConnection(
@@ -114,10 +99,5 @@ router.get('/signup', (req, res) => {
   }
   res.render('signup');
 });
-
-
-
-
-
 
 module.exports = router;
