+++ controllers/index.js
@@ -2,8 +2,10 @@ const router = require('express').Router();
 
 const apiRoutes = require('./api/');
 const homeRoutes = require('./home-routes.js');
+const dashbardRoutes = require('./dashboard-routes.js')
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
+// router.use('/dashbard', dashbardRoutes);
 
 module.exports = router;
