+++ public/js/dashboard.js
@@ -41,7 +41,7 @@ formValue.addEventListener("click", function(clicked) {
     previousCountries.push(val);
     localStorage.setItem("searched Countries", JSON.stringify(previousCountries));  
     // call function to start the API call
-    getToken(val);
+    // getToken(val);
   }
 });
 
@@ -110,69 +110,87 @@ function checkIfValid(value){
 
 
 //calling API to recieve access token from Amadeus
-function getToken(apiCountry){
-  console.log(`starting api call to get token`);
-  //api body parameters
-  var data = 
-    `grant_type=client_credentials&client_id=${CLIENT_ID}&client_secret=${CLIENT_SECRET}`
-  ;
+// const callingAPI1 = async function(event){
+//   event.preventDefault();
+//   console.log(`event made`);
+  // var body1 = 
+  //     `grant_type=client_credentials&client_id=${CLIENT_ID}&client_secret=${CLIENT_SECRET}`
+  //   ;
+  
+  // const response = await fetch('https://test.api.amadeus.com/v1/security/oauth2/token', {
+  //   method: 'POST',
+  //   body: body1,
+  //   headers: {'Content-Type': 'application/x-www-form-urlencoded'}
+  // })
+  // const data = await response.json;
 
-  //Setting up method, url, headers for api Oauth2.0 request
-  var config = {
-    method: 'post',
-    url: 'https://test.api.amadeus.com/v1/security/oauth2/token',
-    headers: { 
-    'Content-Type': 'application/x-www-form-urlencoded'
-    },
-    data : data
-  };
 
-  axios(config)
-    .then(function (response) {
-      console.log(`----____--_-__-----___--`);
-      console.log(JSON.stringify(response.data));
-      //console.log(response.data.access_token)
-      let access_token = response.data.access_token;
-      retrieveCountry(access_token, apiCountry);
-  })
-  .catch(function (error) {
-    console.log(error);
-  });
-}
+
+  // function getToken(apiCountry){
+  //   console.log(`starting api call to get token`);
+  //   //api body parameters
+  //   var data = 
+  //     `grant_type=client_credentials&client_id=${CLIENT_ID}&client_secret=${CLIENT_SECRET}`
+  //   ;
+  
+  //   //Setting up method, url, headers for api Oauth2.0 request
+  //   var config = {
+  //     method: 'post',
+  //     url: 'https://test.api.amadeus.com/v1/security/oauth2/token',
+  //     headers: { 
+  //     'Content-Type': 'application/x-www-form-urlencoded'
+  //     },
+  //     data : data
+  //   };
+  
+  //   axios(config)
+  //     .then(function (response) {
+  //       console.log(`----____--_-__-----___--`);
+  //       console.log(JSON.stringify(response.data));
+  //       //console.log(response.data.access_token)
+  //       let access_token = response.data.access_token;
+  //       retrieveCountry(access_token, apiCountry);
+  //   })
+  //   .catch(function (error) {
+  //     console.log(error);
+  //   });
+  // }
+// }
+
 
 //retrieve country's COVID information
-function retrieveCountry(token, searchCountry) {
-  console.log(`token has been aquired, fetching COVID data`);
-  console.log(token);
-  var data1 = ``;
+// function retrieveCountry(token, searchCountry) {
+//   console.log(`token has been aquired, fetching COVID data`);
+//   console.log(token);
+//   var data1 = ``;
   
-  var config = {
-      method: 'get',
-      url: `https://test.api.amadeus.com/v1/duty-of-care/diseases/covid19-area-report?countryCode=${searchCountry}`,
-      headers: { 
-          'Content-Type': 'application/x-www-form-urlencoded',
-          'Authorization-Bearer': 'G59rXFmdGmc8q0AF2FyN3j85kKVq', 
-          'Authorization': 'Bearer '+ token
-      },
-      data : data1
-  };
+//   var config = {
+//       method: 'get',
+//       url: `https://test.api.amadeus.com/v1/duty-of-care/diseases/covid19-area-report?countryCode=${searchCountry}`,
+//       headers: { 
+//           'Content-Type': 'application/x-www-form-urlencoded',
+//           'Authorization-Bearer': 'G59rXFmdGmc8q0AF2FyN3j85kKVq', 
+//           'Authorization': 'Bearer '+ token
+//       },
+//       data : data1
+//   };
 
-  axios(config)
-  .then(function (response) {
-      let apiData = response.data;
-      // console.log(JSON.stringify(response.data));
-      console.log(apiData);
-      //console.log(apiData.data.area.name);
-      //console.log(apiData.data.diseaseRiskLevel);
-      //console.log(apiData.data.areaAccessRestriction.declarationDocuments);
-      //console.log(apiData.data.areaAccessRestriction.entry.text);
-      //console.log(apiData.data.areaAccessRestriction.mask);
-      setDashboard(apiData);
-  })
-  .catch(function (error) {
-      console.log(error);
-  });
-}
+//   axios(config)
+//   .then(function (response) {
+//       let apiData = response.data;
+//       // console.log(JSON.stringify(response.data));
+//       console.log(apiData);
+//       //console.log(apiData.data.area.name);
+//       //console.log(apiData.data.diseaseRiskLevel);
+//       //console.log(apiData.data.areaAccessRestriction.declarationDocuments);
+//       //console.log(apiData.data.areaAccessRestriction.entry.text);
+//       //console.log(apiData.data.areaAccessRestriction.mask);
+//       setDashboard(apiData);
+//   })
+//   .catch(function (error) {
+//       console.log(error);
+//   });
+// }
 
 // set data to elements to show in dashboard-handlebars
 function setDashboard(covidData) {
