+++ server.js
@@ -38,14 +38,6 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-
-// app.use(
-//   cors({
-//     origin: '*',
-//     methods: ['GET','POST'],
-//     allowedHeaders: ['Content-Type', 'Authorization-Bearer'],
-//     optionsSuccessfulStatue: 204
-// }));
 app.use(routes);
 
 // runs sequelize and checks if there is a database , if not, then it will run the schema and creates
@@ -56,6 +48,3 @@ sequelize.sync({ force: false }).then(() => {
       `\nServer running on port ${PORT}. Visit http://localhost:${PORT}`
     ));
 });
-
-
-
